CalculateInterest

+++ InterestApi/InterestApi/Entities/InterestsOptions.cs
@@ -2,7 +2,7 @@
 {
     public class InterestsOptions
     {
-        public byte InterestRate { get; set; }
+        public double InterestRate { get; set; }
         public string DesiredAmountNullError { get; set; }
         public string DesiredAmountNegativeError { get; set; }
         public string MaturityAmountNullError { get; set; }
